Updated server/main.lua
The current server/main.lua throws a index error for the is statement at 294 (now 297). The changes fix that error by defining the Config values in a local. The config file is linked in your resource.lua so I am not sure why this particular statement threw an error. However it will no longer do so :)